drm/atomic: Boost CPU when committing a new frame
Reimplementation of Sultan's commit but without the devfreq booster. In order to reduce jank, request a CPU boost whenever a new frame is ready to be rendered to the display. The boosts should be sufficient enough to render 60 FPS without any dropped frames when there is no significant external source of load.